+++ NHibernate.Caches/Redis/NHibernate.Caches.Redis/RedisProvider.cs
@@ -46,8 +46,8 @@ namespace NHibernate.Caches.Redis
         private static readonly RedisGarbageCollector GarbageCollector;
 
 	    private static readonly int concurrencyIdBlockSize = 100000;
-	    private static int nextConcurrencyId=-1;
-	    private static int concurrencyIdUpperLimit=-1;
+	    private static long nextConcurrencyId=-1;
+	    private static long concurrencyIdUpperLimit=-1;
 
 	    private static readonly string concurrencyIdKey = RedisNamespace.Uniqueifier + "NHREDIS_CONCURRENCY_IDS";
